Move metadata code to submodule of syncer
